# ips-frontend/src/app/api/models/plan.ts
/* tslint:disable */
export interface Plan {
  id?: number;

  /**
   * Plan identf. name
   */
  planName?: string;

  /**
   * Encoded place image
   */
  planImage?: string;

  /**
   * Plan width
   */
  planWidth?: number;

  /**
   * Plan height
   */
  planHeight?: number;

  /**
   * Plan scale
   */
  planScale?: number;
}